Update base.html, added "Logout" text and symbol when we are logged in and "Profile" option in Navbar when logged in
Updated the base.html so that when the user is logged in it shows "Logout" in the navbar to reflect that. Also changed the "icon-login" font awesome icon to a signout font awesome icon. Added a "Profile" option in the navbar when we are logged in so that the user can navigate to the "/profile" page.